Replaced INT_TO_POINTER and POINTER_TO_INT macros with POINTER_CAST and POINTER_CAST_TO macros, the CAST_TO takes a parameter to which type we're casting to. Also POINTER_CAST should be valid ANSI-C now.
--HG--
branch : HEAD

+++ src/lib-storage/index/index-msgcache.c
@@ -73,7 +73,7 @@ static MessagePart *index_msgcache_get_cached_parts(Pool pool, void *context)
 	IndexMsgcacheContext *ctx = context;
 	MessagePart *part;
 	const void *part_data;
-	unsigned int part_size;
+	size_t part_size;
 
 	part_data = ctx->index->lookup_field_raw(ctx->index, ctx->rec,
 						 FIELD_TYPE_MESSAGEPART,
